formatting

+++ src/rt/lua_api/lua_graphics.c
@@ -5,8 +5,8 @@ int rose_lua_graphics_pset(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     if (nargs == 2) {
-        uint16_t x = lua_tointeger(L, 1);
-        uint16_t y = lua_tointeger(L, 2);
+        uint16_t x = (uint16_t) lua_tointeger(L, 1);
+        uint16_t y = (uint16_t) lua_tointeger(L, 2);
         int err = rose_api_graphics_pset_default(r, x, y);
         switch (err) {
             case ROSE_API_ERR_OUT_OF_BOUNDS:
@@ -16,9 +16,9 @@ int rose_lua_graphics_pset(lua_State* L) {
                 break;
         }
     } else if (nargs >= 3) {
-        uint16_t x = lua_tointeger(L, 1);
-        uint16_t y = lua_tointeger(L, 2);
-        uint8_t val = lua_tointeger(L, 3);
+        uint16_t x = (uint16_t) lua_tointeger(L, 1);
+        uint16_t y = (uint16_t) lua_tointeger(L, 2);
+        uint8_t val = (uint8_t) lua_tointeger(L, 3);
         int err = rose_api_graphics_pset(r, x, y, val);
         switch (err) {
             case ROSE_API_ERR_OUT_OF_BOUNDS:
@@ -36,8 +36,8 @@ int rose_lua_graphics_pget(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     if (nargs >= 2) {
-        uint16_t x = lua_tointeger(L, 1);
-        uint16_t y = lua_tointeger(L, 2);
+        uint16_t x = (uint16_t) lua_tointeger(L, 1);
+        uint16_t y = (uint16_t) lua_tointeger(L, 2);
         uint8_t res;
         int err = rose_api_graphics_pget(r, x, y, &res);
         switch (err) {
@@ -60,10 +60,10 @@ int rose_lua_graphics_palset(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     if (nargs >= 4) {
-        uint8_t idx = lua_tointeger(L, 1);
-        uint8_t red = lua_tointeger(L, 2);
-        uint8_t green = lua_tointeger(L, 3);
-        uint8_t blue = lua_tointeger(L, 4);
+        uint8_t idx = (uint8_t) lua_tointeger(L, 1);
+        uint8_t red = (uint8_t) lua_tointeger(L, 2);
+        uint8_t green = (uint8_t) lua_tointeger(L, 3);
+        uint8_t blue = (uint8_t) lua_tointeger(L, 4);
         Rose_Color c = {.r = red, .g = green, .b = blue};
         int err = rose_api_graphics_palset(r, idx, c);
         switch (err) {
@@ -83,7 +83,7 @@ int rose_lua_graphics_palget(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     if (nargs >= 1) {
-        uint8_t idx = lua_tointeger(L, 1);
+        uint8_t idx = (uint8_t) lua_tointeger(L, 1);
         Rose_Color res;
         int err = rose_api_graphics_palget(r, idx, &res);
         switch (err) {
@@ -106,13 +106,13 @@ int rose_lua_graphics_line(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     int err;
-    int16_t x0 = lua_tointeger(L, 1);
-    int16_t y0 = lua_tointeger(L, 2);
-    int16_t x1 = lua_tointeger(L, 3);
-    int16_t y1 = lua_tointeger(L, 4);
+    int16_t x0 = (int16_t) lua_tointeger(L, 1);
+    int16_t y0 = (int16_t) lua_tointeger(L, 2);
+    int16_t x1 = (int16_t) lua_tointeger(L, 3);
+    int16_t y1 = (int16_t) lua_tointeger(L, 4);
     printf("%hd %hd %hd %hd\n", x0,y0,x1,y1);
     if (nargs >= 5) {
-        uint8_t col = lua_tointeger(L, 5);
+        uint8_t col = (uint8_t) lua_tointeger(L, 5);
         err = rose_api_graphics_line(r, x0, y0, x1, y1, col);
     } else {
         err = rose_api_graphics_line_default(r, x0, y0, x1, y1);
@@ -137,27 +137,27 @@ int rect_lua_common(lua_State* L, rect_fp rect, rect_default_fp rect_default) {
     int nargs = lua_gettop(L);
     int err = ROSE_API_ERR_NONE;
     if (nargs >= 4) {
-        int16_t x0 = lua_tointeger(L, 1);
-        int16_t y0 = lua_tointeger(L, 2);
-        int16_t x1 = lua_tointeger(L, 3);
-        int16_t y1 = lua_tointeger(L, 4);
+        int16_t x0 = (int16_t) lua_tointeger(L, 1);
+        int16_t y0 = (int16_t) lua_tointeger(L, 2);
+        int16_t x1 = (int16_t) lua_tointeger(L, 3);
+        int16_t y1 = (int16_t) lua_tointeger(L, 4);
         if (nargs >= 5) {
-            uint8_t col = lua_tointeger(L, 5);
+            uint8_t col = (uint8_t) lua_tointeger(L, 5);
             err = rect(r, x0, y0, x1, y1, col);
         } else {
             err = rect_default(r, x0, y0, x1, y1);
         }
         
     } else if (nargs == 3) {
-        int16_t x0 = lua_tointeger(L, 1);
-        int16_t y0 = lua_tointeger(L, 2);
-        int16_t x1 = lua_tointeger(L, 3);
+        int16_t x0 = (int16_t) lua_tointeger(L, 1);
+        int16_t y0 = (int16_t) lua_tointeger(L, 2);
+        int16_t x1 = (int16_t) lua_tointeger(L, 3);
         err = rect_default(r, x0, y0, x1, 0);
     } else if (nargs >= 1) {
-        int16_t x1 = lua_tointeger(L, 1);
+        int16_t x1 = (int16_t) lua_tointeger(L, 1);
         int16_t y1 = 0;
         if (nargs == 2) {
-            y1 = lua_tointeger(L, 2);
+            y1 = (int16_t) lua_tointeger(L, 2);
         }
         err = rect_default(r, 0, 0, x1, y1);
     }
@@ -190,16 +190,16 @@ int circ_lua_common(lua_State* L, circ_fp circ, circ_default_fp circ_default) {
     int nargs = lua_gettop(L);
     int err = ROSE_API_ERR_NONE;
     if (nargs >= 2) {
-        int16_t x0 = lua_tointeger(L, 1);
-        int16_t y0 = lua_tointeger(L, 2);
+        int16_t x0 = (int16_t) lua_tointeger(L, 1);
+        int16_t y0 = (int16_t) lua_tointeger(L, 2);
         uint16_t radius;
         if (nargs >= 3) {
-            radius = lua_tointeger(L, 3);
+            radius = (uint16_t) lua_tointeger(L, 3);
         } else {
             radius = 4;
         }
         if (nargs >= 4) {
-            uint8_t col = lua_tointeger(L, 4);
+            uint8_t col = (uint8_t) lua_tointeger(L, 4);
             err = circ(r, x0, y0, radius, col);
         } else {
             err = circ_default(r, x0, y0, radius);
