+++ src/rt/lua_api/lua_memory.c
@@ -3,8 +3,8 @@
 int rose_lua_memory_poke(lua_State* L) {
     Rose_RuntimeBase* r = rose_lua_base_get_runtime_base(L);
     if (r == NULL) { return 0; }
-    uint32_t idx = lua_tointeger(L, 1);
-    uint8_t val = lua_tointeger(L, 2);
+    uint32_t idx = (uint32_t) lua_tointeger(L, 1);
+    uint8_t val = (uint8_t) lua_tointeger(L, 2);
     int err = rose_api_memory_poke(r, idx, val);
     switch (err) {
         case ROSE_API_ERR_OUT_OF_BOUNDS:
@@ -21,7 +21,7 @@ int rose_lua_memory_peek(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     if (nargs >= 1) {
-        uint32_t idx = lua_tointeger(L, 1);
+        uint32_t idx = (uint32_t) lua_tointeger(L, 1);
         uint8_t res;
         int err = rose_api_memory_peek(r, idx, &res);
         switch (err) {
@@ -44,9 +44,9 @@ int rose_lua_memory_memcpy(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     if (nargs >= 3) {
-        uint32_t dest_addr = lua_tointeger(L, 1);
-        uint32_t source_addr = lua_tointeger(L, 2);
-        uint32_t len = lua_tointeger(L, 3);
+        uint32_t dest_addr = (uint32_t) lua_tointeger(L, 1);
+        uint32_t source_addr = (uint32_t) lua_tointeger(L, 2);
+        uint32_t len = (uint32_t) lua_tointeger(L, 3);
         int err = rose_api_memory_memcpy(r, dest_addr, source_addr, len);
         switch (err) {
             case ROSE_API_ERR_OUT_OF_BOUNDS:
@@ -65,9 +65,9 @@ int rose_lua_memory_memset(lua_State* L) {
     if (r == NULL) { return 0; }
     int nargs = lua_gettop(L);
     if (nargs >= 3) {
-        uint32_t dest_addr = lua_tointeger(L, 1);
-        uint8_t val = lua_tointeger(L, 2);
-        uint32_t len = lua_tointeger(L, 3);
+        uint32_t dest_addr = (uint32_t) lua_tointeger(L, 1);
+        uint8_t val = (uint8_t) lua_tointeger(L, 2);
+        uint32_t len = (uint32_t) lua_tointeger(L, 3);
         int err = rose_api_memory_memset(r, dest_addr, val, len);
         switch (err) {
             case ROSE_API_ERR_OUT_OF_BOUNDS:
